Decode regions.
Added a function get_region that returns an Enum representing the region of the
GCI, based on the final character of the Gamecode. It supports Japan, Europe,
and US regions, which are all the regions according to my reference.

I would just implement a three-character Gamecode and have the region be an
entry in the dict, but I want the dict to mirror the format used by dolphin, so
that's not an option.

I'm also not sure if I want to allow setting the region directly, or if I
should just require the user to manually change the final letter of the
Gamecode. Something to decide later.